utils: move add_suffix to sct.image
It deals with naming of image files so it belongs there.
