+++ spinalcordtoolbox/utils.py
@@ -313,24 +313,6 @@ def abspath(fname):
     return os.path.abspath(os.path.expanduser(fname))
 
 
-def add_suffix(fname, suffix):
-    """
-    Add suffix between end of file name and extension.
-
-    :param fname: absolute or relative file name. Example: t2.nii
-    :param suffix: suffix. Example: _mean
-    :return: file name with suffix. Example: t2_mean.nii
-
-    Examples:
-    .. code:: python
-
-        add_suffix(t2.nii, _mean) -> t2_mean.nii
-        add_suffix(t2.nii.gz, a) -> t2a.nii.gz
-    """
-    stem, ext = splitext(fname)
-    return os.path.join(stem + suffix + ext)
-
-
 def check_exe(name):
     """
     Ensure that a program exists
@@ -431,23 +413,6 @@ def parse_num_list_inv(list_int):
             colon_is_present = False
 
     return str_num
-
-
-def splitext(fname):
-    """
-    Split a fname (folder/file + ext) into a folder/file and extension.
-
-    Note: for .nii.gz the extension is understandably .nii.gz, not .gz
-    (``os.path.splitext()`` would want to do the latter, hence the special case).
-    """
-    dir, filename = os.path.split(fname)
-    for special_ext in ['.nii.gz', '.tar.gz']:
-        if filename.endswith(special_ext):
-            stem, ext = filename[:-len(special_ext)], special_ext
-            return os.path.join(dir, stem), ext
-    # If no special case, behaves like the regular splitext
-    stem, ext = os.path.splitext(filename)
-    return os.path.join(dir, stem), ext
 
 
 def tmp_create(basename=None):
